src/requirements.txt
gradio>=4.0.0[mcp]
gradio>=4.0.0
minionx>=0.1.3
python-dotenv>=1.0.0
mcp>=1.0.0
transformers>=4.35.0